extract stats package
