+++ cmd/lotus-stats/setup.bash
