+++ tools/stats/head_buffer.go
@@ -1,4 +1,4 @@
-package main
+package stats
 
 import (
 	"container/list"
