+++ tools/stats/head_buffer_test.go
@@ -1,4 +1,4 @@
-package main
+package stats
 
 import (
 	"testing"
