+++ tools/stats/rpc.go
@@ -1,4 +1,4 @@
-package main
+package stats
 
 import (
 	"context"
